+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
+## 0.12.0 (1. January, 2024)
+### Changed
+- (Breaking) Updated to Axum_session 0.12.0
+- added tracing for debugging.
+- Fixed the return to not be an unwrap if session doesnt exist.
+
 ## 0.11.0 (21. December, 2023)
 ### Changed
 - (Breaking) Updated to Axum_session 0.11.0 for Redis_pool 0.3